korpexport: Support formatting token attrs as list
korpexport: Support formatting the values of an attribute for all the
tokens of a sentence as a single list. The attributes to be (potentially)
formatted are listed in the option sentence_token_attrs, and
sentence_fields may contain references to them. For example, if
sentence_token_attrs is "lemma", sentence_fields may refer to lemmas_all,
lemmas_match, lemmas_left_context and lemmas_right_context.

korpexport/formatter.py:
  KorpExportFormatter:
    _option_defaults: Default values for sentence_token_attrs,
       token_attr_format, attr_only_format.
    Added _init_sentence_token_attrs() initializing _sentence_token_attrs,
      _sentence_token_attr_labels.
    _format_sentence(): Support formatting the attributes in
       _sentence_token_attrs as a single list.
    _format_token(): Support option attr_only for formatting a bare
      attribute.
    _format_token_attr(): Support specifying a different format via the
      keyword argument item_type.