korpexport: Support subformats as shorthands
korpexport: Support subformats as a kind of a shorthand for groups of
default option values for a formatting class, without having to add a
subclass. A formatting class specifies subformats in the class attribute
_subformat_options, which is a dict whose keys are subformat names and
values the corresponding option defaults. A subformat is used by specifying
the subformat name(s) in the URL parameter "subformat" (a comma-separated
list).

korpexport/exporter.py:
  Pass the subformat parameter from the input form to the formatter
    constructor. (KorpExporter._get_formatter)

korpexport/formatter.py:
  KorpExportFormatter:
    Added (empty) class attribute _subformats.
    __init__(): Apply the option defaults from all the specified
      subformats.